+++ backend/src/main/java/com/patriot/fourlipsclover/member/repository/MemberRepository.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 
     Optional<Member> findByEmail(String email);
 
-    Member findByMemberId(long l);
+    Member findByMemberId(Long l);
 
 }
